requirements.txt
faiss-cpu
groq
langchain-groq
pyPDF2
langchain
langchain_google_genai
streamlit
langchain_community
python-dotenv
pypdf
PyMuPDFb-1.24.6 
docx2txt-0.8 
pymupdf-1.24.7
